Moved the definition and functions related to positions to a different header.
